feat: Adding Jest Configurator to Gradient package
