+++ packages/gradient/__tests__/android/gradient.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Gradient tests on android", () => {
+  shared();
+});
