+++ packages/gradient/__tests__/ios/gradient.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Gradient tests on iOS", () => {
+  shared();
+});
