+++ packages/gradient/__tests__/web/gradient.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Gradient tests on web", () => {
+  shared();
+});
