+++ packages/gradient/__tests__/web/jest.config.js
@@ -1,0 +1,3 @@
+const jestConfigurator = require("@times-components/jest-configurator").default;
+
+module.exports = jestConfigurator("web", __dirname);
