+++ packages/gradient/package.json
@@ -4,8 +4,11 @@
   "description": "Renders a gradient background",
   "main": "gradient",
   "scripts": {
-    "test:watch": "jest --bail --verbose --watchAll",
-    "test": "jest --bail --ci --coverage",
+    "test:android": "jest --config='./__tests__/android/jest.config.js'",
+    "test:ios": "jest --config='./__tests__/ios/jest.config.js'",
+    "test:web": "jest --config='./__tests__/web/jest.config.js'",
+    "test": "yarn test:android --ci --bail --coverage && yarn test:ios --ci --bail --coverage && yarn test:web --ci --bail --coverage",
+    "test:watch": "jest --projects */**/jest.config.js --watchAll",
     "fmt": "prettier --write '**/*.*'",
     "prettier:diff": "prettier --list-different '**/*.*'",
     "lint": "eslint . && yarn prettier:diff",
@@ -54,6 +57,7 @@
     "react-test-renderer": "16.2.0"
   },
   "dependencies": {
+    "@times-components/jest-configurator": "0.0.22",
     "prop-types": "15.6.0",
     "react-native-linear-gradient": "2.3.0"
   },
